added multi-iteration test
I personally tested this with 1000 iterations many times and it works
fine.  In the interest of not running too slowly and/or breaking out a
"slow" suite of tests, I dailed the iterations back to 50 which is
relatively quick while still proving the point.

Also added count of the number of stored ReceivableTokens just to
provide assurance that the storage gets curated properly.